# Big integer backend that executes GMP's constant-time arithmetic functions when possible