--- a/CITATION.cff
+++ b/CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Jelle"
   orcid: "https://orcid.org/0000-0002-3979-9740"
 title: "scicrypt"
-version: 0.6.2
+version: 0.7.0
 date-released: 2022-09-08
 url: "https://github.com/jellevos/scicrypt"