--- a/scicrypt-bigint/src/arithmetic/add.rs
+++ b/scicrypt-bigint/src/arithmetic/add.rs
@@ -2,10 +2,12 @@ use std::{ops::{AddAssign, Add}, cmp::{min, max}, iter::Sum};
 
 use gmp_mpfr_sys::gmp;
 
-use crate::BigInteger;
+use crate::{BigInteger, scratch::Scratch, GMP_NUMB_BITS};
 
 impl AddAssign<&BigInteger> for BigInteger {
     fn add_assign(&mut self, rhs: &Self) {
+        debug_assert!(self.size_in_bits >= rhs.size_in_bits);
+
         let n = min(self.value.size, rhs.value.size);
 
         if n == 0 {
@@ -37,10 +39,68 @@ impl Add<&BigInteger> for BigInteger {
     }
 }
 
-// TODO: Also implement addition with u64 using `mpn_sec_add_1`
+impl AddAssign<u64> for BigInteger {
+    fn add_assign(&mut self, rhs: u64) {
+        unsafe {
+            let scratch_size = gmp::mpn_sec_add_1_itch(self.value.size as i64)
+                as usize
+                * GMP_NUMB_BITS as usize;
+            
+            let mut scratch = Scratch::new(scratch_size);
+
+            let carry = gmp::mpn_sec_add_1(
+                self.value.d.as_mut(),
+                self.value.d.as_ptr(),
+                self.value.size as i64,
+                rhs,
+                scratch.as_mut()
+            );
+
+            self.value.size += carry as i32;
+            self.size_in_bits += carry as i64;
+        }
+    }
+}
+
+impl Add<u64> for BigInteger {
+    type Output = BigInteger;
+
+    fn add(mut self, rhs: u64) -> Self::Output {
+        self += rhs;
+        self
+    }
+}
 
 impl<'a> Sum<&'a BigInteger> for BigInteger {
-    fn sum<I: Iterator<Item = &'a BigInteger>>(iter: I) -> Self {
-        todo!()
+    fn sum<I: Iterator<Item = &'a BigInteger>>(mut iter: I) -> Self {
+        let initial = iter.next().unwrap().clone();
+        iter.fold(initial, |x, y| x + y)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::BigInteger;
+
+    #[test]
+    fn test_addition() {
+        let mut x = BigInteger::from_string("5378239758327583290580573280735".to_string(), 10, 103);
+        let y = BigInteger::from_string("49127277414859531000011129".to_string(), 10, 86);
+
+        x += &y;
+
+        assert_eq!(BigInteger::from_string("5378288885604998150111573291864".to_string(), 10, 103), x);
+        assert_eq!(x.size_in_bits, 103);
+    }
+
+    #[test]
+    fn test_addition_u64() {
+        let mut x = BigInteger::from_string("5378239758327583290580573280735".to_string(), 10, 103);
+        let y = 14;
+
+        x += y;
+
+        assert_eq!(BigInteger::from_string("5378239758327583290580573280749".to_string(), 10, 103), x);
+        assert_eq!(x.size_in_bits, 103);
     }
 }