--- a/scicrypt-bigint/src/arithmetic/div.rs
+++ b/scicrypt-bigint/src/arithmetic/div.rs
@@ -1,0 +1,9 @@
+use std::ops::DivAssign;
+
+use crate::BigInteger;
+
+impl DivAssign<&BigInteger> for BigInteger {
+    fn div_assign(&mut self, rhs: &BigInteger) {
+        todo!()
+    }
+}