--- a/scicrypt-bigint/src/arithmetic/div.rs
+++ b/scicrypt-bigint/src/arithmetic/div.rs
@@ -1,6 +1,6 @@
 use std::ops::{DivAssign, Div};
 
-use gmp_mpfr_sys::gmp;
+use gmp_mpfr_sys::{gmp};
 
 use crate::{BigInteger, GMP_NUMB_BITS, scratch::Scratch};
 
@@ -9,8 +9,11 @@ impl Div<&BigInteger> for BigInteger {
 
     fn div(mut self, rhs: &BigInteger) -> BigInteger {
         // TODO: Check the other preconditions
-        debug_assert_eq!(self.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, self.value.size, "the operands' size in bits must match their actual size");
-        debug_assert_eq!(rhs.size_in_bits.div_ceil(GMP_NUMB_BITS as i64) as i32, rhs.value.size, "the operands' size in bits must match their actual size");
+        debug_assert_eq!(self.size_in_bits.div_ceil(GMP_NUMB_BITS) as i32, self.value.size.abs(), "the operands' size in bits must match their actual size");
+        debug_assert_eq!(rhs.size_in_bits.div_ceil(GMP_NUMB_BITS) as i32, rhs.value.size.abs(), "the operands' size in bits must match their actual size");
+
+        debug_assert!(self.value.size.abs() >= rhs.value.size.abs());
+        debug_assert!(rhs.value.size.abs() >= 1);
 
         unsafe {
             let scratch_size = gmp::mpn_sec_div_qr_itch(self.value.size as i64, rhs.value.size as i64)
@@ -19,35 +22,27 @@ impl Div<&BigInteger> for BigInteger {
 
             let mut scratch = Scratch::new(scratch_size);
 
-            let mut res = BigInteger::init(self.value.size - rhs.value.size);
+            let mut res = BigInteger::init(self.value.size.abs() - rhs.value.size.abs() + 1);
 
-            gmp::mpn_sec_div_qr(
+            let most_significant_limb = gmp::mpn_sec_div_qr(
                 res.value.d.as_mut(),
                 self.value.d.as_mut(),
-                self.value.size as i64,
+                self.value.size.abs() as i64,
                 rhs.value.d.as_ptr(),
-                rhs.value.size as i64,
+                rhs.value.size.abs() as i64,
                 scratch.as_mut(),
             );
 
-            res.value.size = self.value.size - rhs.value.size + 1;
-            res.size_in_bits = self.size_in_bits - rhs.size_in_bits + 1;
+            // FIXME: Logic for negative size
+            println!("MSL: {}", most_significant_limb);
+            let sign = self.value.size.signum() * rhs.value.size.signum();
+            let size = self.value.size.abs() - rhs.value.size.abs();
+            res.value.size = sign * (size + (most_significant_limb != 0) as i32);
+            //res.size_in_bits = self.size_in_bits - rhs.size_in_bits + 1;
+            res.size_in_bits = res.value.size.abs() as u32 * GMP_NUMB_BITS;
+            res.value.d.as_ptr().offset(size as isize).write(most_significant_limb);
             return res;
         }
-    }
-}
-
-impl Div<i64> for BigInteger {
-    type Output = BigInteger;
-
-    fn div(mut self, rhs: i64) -> Self::Output {
-        self = self / &BigInteger::from(rhs.abs() as u64);
-
-        if rhs.is_negative() {
-            self.value.size = -self.value.size;
-        }
-
-        self
     }
 }
 
@@ -56,13 +51,63 @@ mod test {
     use crate::BigInteger;
 
     #[test]
+    fn test_division_small() {
+        let x = BigInteger::from_string("5".to_string(), 10, 3);
+        let y = BigInteger::from_string("3".to_string(), 10, 2);
+
+        dbg!(&x);
+        dbg!(&y);
+
+        let q = x / &y;
+        dbg!(&q);
+
+        assert_eq!(BigInteger::from_string("1".to_string(), 10, 1), q);
+        assert_eq!(q.value.size, 1);
+        assert_eq!(q.size_in_bits, 64);
+    }
+
+    #[test]
+    fn test_division_small_zero() {
+        let x = BigInteger::from_string("4".to_string(), 10, 3);
+        let y = BigInteger::from_string("7".to_string(), 10, 3);
+
+        dbg!(&x);
+        dbg!(&y);
+
+        let q = x / &y;
+        dbg!(&q);
+
+        assert_eq!(BigInteger::from_string("0".to_string(), 10, 1), q);
+        assert_eq!(q.value.size, 0);
+        assert_eq!(q.size_in_bits, 0);
+    }
+
+    #[test]
     fn test_division() {
         let x = BigInteger::from_string("5378239758327583290580573280735".to_string(), 10, 103);
         let y = BigInteger::from_string("49127277414859531000011129".to_string(), 10, 86);
 
+        dbg!(&x);
+        dbg!(&y);
+
         let q = x / &y;
+        dbg!(&q);
 
         assert_eq!(BigInteger::from_string("109475".to_string(), 10, 17), q);
-        assert_eq!(q.size_in_bits, 17);
+        assert_eq!(q.value.size, 1);
+        assert_eq!(q.size_in_bits, 64);
+    }
+
+    #[test]
+    fn test_division_negative() {
+        let x = BigInteger::from_string("5378239758327583290580573280735".to_string(), 10, 103);
+        let y = BigInteger::from_string("-49127277414859531000011129".to_string(), 10, 86);
+
+        let q = x / &y;
+        dbg!(&q);
+
+        assert_eq!(BigInteger::from_string("-109475".to_string(), 10, 17), q);
+        assert_eq!(q.value.size, -1);
+        assert_eq!(q.size_in_bits, 64);
     }
 }