--- a/scicrypt-bigint/src/arithmetic/mod.rs
+++ b/scicrypt-bigint/src/arithmetic/mod.rs
@@ -1,0 +1,4 @@
+mod add;
+mod sub;
+mod mul;
+mod div;