--- a/scicrypt-bigint/src/modular/mod.rs
+++ b/scicrypt-bigint/src/modular/mod.rs
@@ -1,0 +1,3 @@
+mod rem;
+mod pow;
+mod inv;