This PR adds a new sub-crate to scicrypt that implements many functions from rug but in a constant amount of time (for constant-size operands). All other sub-crates now use this backend.

**Failing tests:**
- [x] 2-of-3 Paillier
- [x] 2-of-3 integer-based ElGamal
- [x] 3-of-3 integer-based ElGamal (_flaky_)