@jellevos 2022-09-08
Just now I looked through the benchmark from CI and it does seem like a small regression in run time (I have no idea why CI didn't pick that up). However, given that the previous code was in many places completely leaking the size of the input I think that it is acceptable. The biggest slow-downs come from RSA encryption (57 us to 100 us) and integer-ElGamal decryption (15 ms to 16 ms). Both I deem very acceptable.

*Before*
paillier/encrypt        time:   [57.592 ms 57.649 ms 57.741 ms]
paillier/decrypt        time:   [56.520 ms 56.531 ms 56.545 ms]
rsa/encrypt             time:   [57.189 us 57.199 us 57.209 us]
rsa/decrypt             time:   [14.696 ms 14.705 ms 14.718 ms]
curve_elgamal/encrypt   time:   [36.120 us 36.123 us 36.128 us]
curve_elgamal/decrypt   time:   [64.347 us 64.353 us 64.360 us]
integer_elgamal/encrypt time:   [29.520 ms 29.531 ms 29.542 ms]
integer_elgamal/decrypt time:   [14.689 ms 14.691 ms 14.693 ms]

*After*
paillier/encrypt        time:   [57.408 ms 57.416 ms 57.424 ms]
paillier/decrypt        time:   [56.579 ms 56.583 ms 56.588 ms]
rsa/encrypt             time:   [100.48 us 100.49 us 100.50 us]
rsa/decrypt             time:   [14.686 ms 14.687 ms 14.689 ms]
curve_elgamal/encrypt   time:   [36.238 us 36.242 us 36.247 us]
curve_elgamal/decrypt   time:   [64.312 us 64.320 us 64.332 us]
integer_elgamal/encrypt time:   [29.435 ms 29.439 ms 29.444 ms]
integer_elgamal/decrypt time:   [16.327 ms 16.328 ms 16.330 ms]