@jellevos 2022-09-08
This PR solves #31